requirements.txt
Django~=4.2.12
djangorestframework~=3.15.1
pandas~=2.2.2